Magic mode: return Perl's interpretation of the value instead of the literal (#4)
Previously, magic mode meant qq/"-quoted strings behaved like q/'-quoted strings
because it returned the value of the first argument to the overload function
rather than the second.

From the 'overload' perldoc:

The corresponding values are references to functions which take three arguments:
the first one is the initial string form of the constant, the second one is how
Perl interprets this constant, the third one is how the constant is used. Note
that the initial string form does not contain string delimiters, and has
backslashes in backslash-delimiter combinations stripped (thus the value of
delimiter is not relevant for processing of this string). The return value of
this function is how this constant is going to be interpreted by Perl.